test: ✅ split percy snapshot & upload jobs
The main goal here is to be able to cache CT & E2E.

Percy's DOM serialization's resource discovery needs percy local server
This makes it hard to grab the dom snapshot and write it to a file.
We made our own command to fix this.

Cf. https://github.com/percy/cli/issues/1156